# kubernetes/swagger-config copy.yaml
openapi: 3.0.3
info:
  title: Messenger API
  version: 1.0.1
  description: API мессенджера
servers:
  - url: /api
    description: REST API сервера
  - url: ws://localhost:8083
    description: WebSocket сервер для звонков
security:
  - BearerAuth: []
...
paths:
...
  /v1/messages/{messageId}:
    put:
      tags:
        - Сообщения
      summary: Редактировать сообщение
      security:
        - BearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
          description: ID сообщения
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMessageRequest'
      responses:
        '200':
          description: Сообщение отредактировано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    delete:
      tags:
        - Сообщения
      summary: Удалить сообщение
      security:
        - BearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID сообщения
      responses:
        '200':
          description: Сообщение удалено
  /v1/messages/{messageId}/reactions:
    post:
      tags:
        - Сообщения
      summary: Добавить реакцию к сообщению
      security:
        - BearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
          description: ID сообщения
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddReactionRequest'
      responses:
        '200':
          description: Реакция добавлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    delete:
      tags:
        - Сообщения
      summary: Удалить реакцию с сообщения
      security:
        - BearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID сообщения
        - name: emoji
          in: query
          required: true
          schema:
            type: string
          description: Эмодзи реакции
      responses:
        '200':
          description: Реакция удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  # Call Service
...
  /v1/channels:
    get:
      tags:
        - Каналы
      summary: Получить список каналов
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Список каналов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Channel'
    post:
      tags:
        - Каналы
      summary: Создать новый канал
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                description:
                  type: string
                isPrivate:
                  type: boolean
      responses:
        '201':
          description: Канал создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
  /v1/channels/{channelId}:
    get:
      tags:
        - Каналы
      summary: Получить информацию о канале
      security:
        - BearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
          description: ID канала
      responses:
        '200':
          description: Информация о канале
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
    put:
      tags:
        - Каналы
      summary: Обновить информацию о канале
      security:
        - BearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID канала
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                isPrivate:
                  type: boolean
      responses:
        '200':
          description: Канал обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
    delete:
      tags:
        - Каналы
      summary: Удалить канал
      security:
        - BearerAuth: []
      parameters:
        - name: channelId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID канала
      responses:
        '200':
          description: Канал удален
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
          description: ID сообщения
        content:
          type: string
        senderId:
          type: integer
          description: ID отправителя
        attachments:
          type: array
          items:
            type: string
        replyToId:
          type: integer
          description: ID сообщения, на которое отвечают
          nullable: true
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/Reaction'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    EditMessageRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
        attachments:
          type: array
          items:
            type: string
    AddReactionRequest:
      type: object
      required:
        - emoji
      properties:
        emoji:
          type: string
          example: "👍"
    Reaction:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        emoji:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Channel:
      type: object
      properties:
        id:
          type: integer
          description: ID канала
        name:
          type: string
        description:
          type: string
        isPrivate:
          type: boolean
        creatorId:
          type: integer
          description: ID создателя канала
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: integer
          description: ID пользователя
        username:
          type: string
        email:
          type: string
        avatar:
          type: string
        status:
          type: string
          enum: [online, offline]
...